# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API blog service.",
        "title": "Go Blog API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/post/{id}/": {
            "get": {
                "description": "getting post instance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Getting post instance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Post"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Post": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Post author",
                    "type": "integer",
                    "example": 234
                },
                "body": {
                    "description": "Post text",
                    "type": "string",
                    "example": "this is my first psot and ..."
                },
                "id": {
                    "description": "The UUID of a post",
                    "type": "integer",
                    "example": 345
                },
                "title": {
                    "description": "Post title",
                    "type": "string",
                    "example": "first post"
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "JWT",
            "in": "header"
        }
    }
}